feature: Add card information text in Card
