+++ src/components/CardImage/CardImage.css
@@ -1,0 +1,8 @@
+.CardImage {
+  width: 100%;
+  min-height: 19.2rem;
+  height: 19.2rem;
+  border-radius: 1.5rem 1.5rem 0 0;
+  background-position: center;
+  background-size: 100%;
+}
