+++ src/components/CardList/CardList.css
@@ -15,12 +15,3 @@
   box-shadow: 0 0.5rem 2.5rem 0 rgba(0, 0, 0, 0.08);
   border-radius: 1.5rem;
 }
-
-.CardImage {
-  width: 100%;
-  min-height: 19.2rem;
-  height: 19.2rem;
-  border-radius: 1.5rem 1.5rem 0 0;
-  background-position: center;
-  background-size: 100%;
-}
